Move congestion control code to allow other algorithms
Use keyword arguments wherever possible to make it explicit what is
being passed around.
