+++ tests/test_recovery.py
@@ -1,20 +1,16 @@
-import math
 from unittest import TestCase
 
-from aioquic import tls
-from aioquic.quic.packet import PACKET_TYPE_INITIAL, PACKET_TYPE_ONE_RTT
-from aioquic.quic.packet_builder import QuicSentPacket
-from aioquic.quic.rangeset import RangeSet
-from aioquic.quic.recovery import (
-    QuicPacketPacer,
-    QuicPacketRecovery,
-    QuicPacketSpace,
-    QuicRttMonitor,
-)
+from aioquic.quic.congestion.base import QuicRttMonitor, create_congestion_control
+from aioquic.quic.recovery import QuicPacketPacer
 
 
-def send_probe():
-    pass
+class QuicCongestionControlTest(TestCase):
+    def test_create_unknown_congestion_control(self):
+        with self.assertRaises(Exception) as cm:
+            create_congestion_control("bogus", max_datagram_size=1280)
+        self.assertEqual(
+            str(cm.exception), "Unknown congestion control algorithm: bogus"
+        )
 
 
 class QuicPacketPacerTest(TestCase):
@@ -59,117 +55,6 @@ class QuicPacketPacerTest(TestCase):
             self.assertIsNone(self.pacer.next_send_time(now=1.00015))
             self.pacer.update_after_send(now=1.00015)
         self.assertAlmostEqual(self.pacer.next_send_time(now=1.00015), 1.0002)
-
-
-class QuicPacketRecoveryTest(TestCase):
-    def setUp(self):
-        self.INITIAL_SPACE = QuicPacketSpace()
-        self.HANDSHAKE_SPACE = QuicPacketSpace()
-        self.ONE_RTT_SPACE = QuicPacketSpace()
-
-        self.recovery = QuicPacketRecovery(
-            initial_rtt=0.1,
-            max_datagram_size=1280,
-            peer_completed_address_validation=True,
-            send_probe=send_probe,
-        )
-        self.recovery.spaces = [
-            self.INITIAL_SPACE,
-            self.HANDSHAKE_SPACE,
-            self.ONE_RTT_SPACE,
-        ]
-
-    def test_discard_space(self):
-        self.recovery.discard_space(self.INITIAL_SPACE)
-
-    def test_on_ack_received_ack_eliciting(self):
-        packet = QuicSentPacket(
-            epoch=tls.Epoch.ONE_RTT,
-            in_flight=True,
-            is_ack_eliciting=True,
-            is_crypto_packet=False,
-            packet_number=0,
-            packet_type=PACKET_TYPE_ONE_RTT,
-            sent_bytes=1280,
-            sent_time=0.0,
-        )
-        space = self.ONE_RTT_SPACE
-
-        #  packet sent
-        self.recovery.on_packet_sent(packet, space)
-        self.assertEqual(self.recovery.bytes_in_flight, 1280)
-        self.assertEqual(space.ack_eliciting_in_flight, 1)
-        self.assertEqual(len(space.sent_packets), 1)
-
-        # packet ack'd
-        self.recovery.on_ack_received(
-            space, ack_rangeset=RangeSet([range(0, 1)]), ack_delay=0.0, now=10.0
-        )
-        self.assertEqual(self.recovery.bytes_in_flight, 0)
-        self.assertEqual(space.ack_eliciting_in_flight, 0)
-        self.assertEqual(len(space.sent_packets), 0)
-
-        # check RTT
-        self.assertTrue(self.recovery._rtt_initialized)
-        self.assertEqual(self.recovery._rtt_latest, 10.0)
-        self.assertEqual(self.recovery._rtt_min, 10.0)
-        self.assertEqual(self.recovery._rtt_smoothed, 10.0)
-
-    def test_on_ack_received_non_ack_eliciting(self):
-        packet = QuicSentPacket(
-            epoch=tls.Epoch.ONE_RTT,
-            in_flight=True,
-            is_ack_eliciting=False,
-            is_crypto_packet=False,
-            packet_number=0,
-            packet_type=PACKET_TYPE_ONE_RTT,
-            sent_bytes=1280,
-            sent_time=123.45,
-        )
-        space = self.ONE_RTT_SPACE
-
-        #  packet sent
-        self.recovery.on_packet_sent(packet, space)
-        self.assertEqual(self.recovery.bytes_in_flight, 1280)
-        self.assertEqual(space.ack_eliciting_in_flight, 0)
-        self.assertEqual(len(space.sent_packets), 1)
-
-        # packet ack'd
-        self.recovery.on_ack_received(
-            space, ack_rangeset=RangeSet([range(0, 1)]), ack_delay=0.0, now=10.0
-        )
-        self.assertEqual(self.recovery.bytes_in_flight, 0)
-        self.assertEqual(space.ack_eliciting_in_flight, 0)
-        self.assertEqual(len(space.sent_packets), 0)
-
-        # check RTT
-        self.assertFalse(self.recovery._rtt_initialized)
-        self.assertEqual(self.recovery._rtt_latest, 0.0)
-        self.assertEqual(self.recovery._rtt_min, math.inf)
-        self.assertEqual(self.recovery._rtt_smoothed, 0.0)
-
-    def test_on_packet_lost_crypto(self):
-        packet = QuicSentPacket(
-            epoch=tls.Epoch.INITIAL,
-            in_flight=True,
-            is_ack_eliciting=True,
-            is_crypto_packet=True,
-            packet_number=0,
-            packet_type=PACKET_TYPE_INITIAL,
-            sent_bytes=1280,
-            sent_time=0.0,
-        )
-        space = self.INITIAL_SPACE
-
-        self.recovery.on_packet_sent(packet, space)
-        self.assertEqual(self.recovery.bytes_in_flight, 1280)
-        self.assertEqual(space.ack_eliciting_in_flight, 1)
-        self.assertEqual(len(space.sent_packets), 1)
-
-        self.recovery._detect_loss(space, now=1.0)
-        self.assertEqual(self.recovery.bytes_in_flight, 0)
-        self.assertEqual(space.ack_eliciting_in_flight, 0)
-        self.assertEqual(len(space.sent_packets), 0)
 
 
 class QuicRttMonitorTest(TestCase):
